backend/requirements.txt
asgiref==3.2.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
defusedxml==0.7.1
Django==3.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
psycopg2-binary==2.8.6
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.3.1
testfixtures==6.18.5
uritemplate==4.1.1
urllib3==1.26.9
gunicorn==20.1.0
drf-base64==2.0
Pillow==9.1.0
fpdf2==2.5.4